Navbar and Cart Button Redirect
Completed navbar for all pages, uniform for every page. Implemented button functionality to redirect to cart.html.

- created cart.html page

- began working on html/css for cart container.

- made each navbar uniform for every page and added script.

TODO:

- create one static cart item and template that html for every item in the cart, similar to products page.

Logic:

Will need to populate cart from object array that has a new product object appended every time user clicks "add to cart". Will probably write functions to handle this on shopPage.js. As user clicks "add to cart" button, product object will be added to an array and then be added to the cart upon loading the cart.html page.

cart.js will handle the actual dynamic populating of the list while shopPage.js will handle adding products to the array that will be utilized by cart.js.

scripting should not be hard to implement, but the css/html is already giving me a headache. No API requests should be needed for this step as the shopPage.js already will have the product information requested on page load. Based on the product id associated with the product card, the add to cart button will save the necessary product information to be repopulated on the cart page.